% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-data.R
\name{augment_cycle_data}
\alias{augment_cycle_data}
\alias{augment_year_data_with_month_resolution}
\alias{augment_year_data_with_second_resolution}
\title{Calculates variables necessary for WATS Plots}
\usage{
augment_year_data_with_month_resolution( dsLinear, dateName )
augment_year_data_with_second_resolution( dsLinear, dateName )
}
\arguments{
\item{dsLinear}{The \link{data.frame} to containing the detailed data.}

\item{dateName}{The variable name in \code{dsLinear} containing the date or datetime value.}
}
\value{
Returns a \link{data.frame} with additional variables: \code{CycleTally}, \code{ProportionThroughCycle}, \code{ProportionID}, and \code{TerminalPointInCycle}.
}
\description{
Calculates variables necessary for WATS Plots.  This the first of two functions
that needs to be called to produce WATS Plots.  \code{\link[=annotate_data]{annotate_data()}} is the second.
}
\examples{
library(Wats)
dsLinear <- county_month_birth_rate_2005_version
dsLinear <- dsLinear[dsLinear$CountyName=="oklahoma", ]
dsLinear <- augment_year_data_with_month_resolution(dsLinear=dsLinear, dateName="Date")
head(dsLinear)

}
