% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian-rolling.R
\name{cartesian_rolling}
\alias{cartesian_rolling}
\title{Linear Plot with Rolling Summaries}
\usage{
cartesian_rolling(
  ds_linear,
  x_name,
  y_name,
  stage_id_name,
  rollingLowerName = "RollingLower",
  rollingCenterName = "RollingCenter",
  rollingUpperName = "RollingUpper",
  palette_dark = NULL,
  palette_light = NULL,
  colorSparse = grDevices::adjustcolor("tan1", 0.5),
  change_points = NULL,
  change_point_labels = NULL,
  drawJaggedLine = TRUE,
  drawRollingLine = TRUE,
  drawRollingBand = TRUE,
  drawSparseLineAndPoints = TRUE,
  jagged_point_size = 2,
  jagged_line_size = 0.5,
  rollingLineSize = 1,
  sparsePointSize = 4,
  sparseLineSize = 0.5,
  bandAlpha = 0.4,
  change_line_alpha = 0.5,
  change_line_size = 3,
  title = NULL,
  x_title = NULL,
  y_title = NULL
)
}
\arguments{
\item{ds_linear}{The \link{data.frame} to containing the data.}

\item{x_name}{The variable name containing the date.}

\item{y_name}{The variable name containing the dependent/criterion variable.}

\item{stage_id_name}{The variable name indicating which stage the record belongs to.  For example, before the first interruption, the \code{StageID} is \code{1}, and is \code{2} afterwards.}

\item{rollingLowerName}{The variable name showing the lower bound of the rolling estimate.}

\item{rollingCenterName}{The variable name showing the rolling estimate.}

\item{rollingUpperName}{The variable name showing the upper bound of the rolling estimate.}

\item{palette_dark}{A vector of colors used for the dark/heavy graphical elements.  The vector should have one color for each \code{StageID} value.  If no vector is specified, a default will be chosen, based on the number of stages.}

\item{palette_light}{A vector of colors used for the light graphical elements.  The vector should have one color for each \code{StageID} value.  If no vector is specified, a default will be chosen, based on the number of stages.}

\item{colorSparse}{The color of the `slowest' trend line, which plots only one value per cycle.}

\item{change_points}{A vector of values indicate the interruptions between stages.  It typically works best as a Date or a POSIXct class.}

\item{change_point_labels}{The text plotted above each interruption.}

\item{drawJaggedLine}{A boolean value indicating if a line should be plotted that connects the observed data points.}

\item{drawRollingLine}{A boolean value indicating if a line should be plotted that connects the rolling estimates specified by \code{rollingCenterName}.}

\item{drawRollingBand}{A boolean value indicating if a band should be plotted that envelopes the rolling estimates (whose values are take from the \code{rollingLowerName} and \code{rollingUpperName}.}

\item{drawSparseLineAndPoints}{A boolean value indicating if the sparse line and points should be plotted.}

\item{jagged_point_size}{The size of the observed data points.}

\item{jagged_line_size}{The size of the line connecting the observed data points.}

\item{rollingLineSize}{The size of the line connecting the rolling estimates.}

\item{sparsePointSize}{The size of the sparse estimates.}

\item{sparseLineSize}{The size of the line connecting the sparse estimates.}

\item{bandAlpha}{The amount of transparency of the rolling estimate band.}

\item{change_line_alpha}{The amount of transparency marking each interruption.}

\item{change_line_size}{The width of a line marking an interruption.}

\item{title}{The string describing the plot.}

\item{x_title}{The string describing the \emph{x}-axis.}

\item{y_title}{The string describing the \emph{y}-axis.}
}
\value{
Returns a ggplot2 graphing object
}
\description{
Shows the interrupted time series in Cartesian coordinates without a periodic/cyclic components.
}
\examples{
library(Wats) #Load the package
changeMonth <- base::as.Date("1996-02-15")
ds_linear <- county_month_birth_rate_2005_version
ds_linear <- ds_linear[ds_linear$CountyName=="oklahoma", ]
ds_linear <- augment_year_data_with_month_resolution(ds_linear=ds_linear, date_name="Date")
hSpread <- function( scores ) { return( quantile(x=scores, probs=c(.25, .75)) ) }
portfolio <- annotate_data(
    ds_linear,
    dv_name = "BirthRate",
    center_function = median,
    spread_function = hSpread
)

cartesian_rolling(
    portfolio$ds_linear,
    x_name = "Date",
    y_name = "BirthRate",
    stage_id_name = "StageID",
    change_points = changeMonth,
    change_point_labels = "Bombing Effect"
)
}
\keyword{Cartesian}
