% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wats-package.R
\docType{package}
\name{Wats-package}
\alias{Wats-package}
\alias{Wats}
\title{Wrap Around Time Series graphics}
\description{
Wrap-around Time Series (WATS) Plots for Interrupted Time Series Designs with Seasonal Patterns
}
\note{
The release version is available through
\href{https://cran.r-project.org/package=Wats}{CRAN} by
running \code{install.packages('Wats')}.
The most recent development version is available through
\href{https://github.com/OuhscBbmc/Wats}{GitHub} by running
\code{remotes::install_github("OuhscBbmc/Wats")}.
(make sure \href{https://cran.r-project.org/package=remotes}{remotes} is already installed).
If you're having trouble with the package, please install the development version.
If this doesn't solve your problem,
please create an \href{https://github.com/OuhscBbmc/Wats/issues}{issue},
or email Will.

\strong{Funding}
Continued development and maintenance of this package supported by the
Oklahoma Shared Clinical and Translational Resources (\href{https://osctr.ouhsc.edu/}{OSCTR},
U54GM104938) with an Institutional Development Award (IDeA) from NIGMS.
The content is solely the responsibility of the authors and does not necessarily represent
the official views of the National Institutes of Health.
}
\examples{
\dontrun{
# Install/update Wats with the release version from CRAN.
install.packages("Wats")

# Install/update Wats with the development version from GitHub
# install.packages("remotes") #Uncomment if remotes isn't installed already.
remotes::install_github("OuhscBbmc/Wats")
}
}
\references{
Rodgers, J.L., Beasley, W.H., and Schuelke, M. (2014).
\href{https://www.tandfonline.com/doi/abs/10.1080/00273171.2014.946589}{Graphical Data Analysis on the Circle:  Wrap-around Time Series Plots for (Interrupted) Time Series Designs.}
\emph{Multivariate Behavioral Research}.

Rodgers, J.L., St. John, C. A. & Coleman R. (2005).
\href{https://pubmed.ncbi.nlm.nih.gov/16463916/}{Did Fertility Go Up after the Oklahoma City Bombing?  An Analysis of Births in Metropolitan Counties in Oklahoma, 1990-1999.}
\emph{Demography, 42}, 675-692.
}
\author{
\href{http://scholar.google.com/citations?user=ffsJTC0AAAAJ}{William Howard Beasley}
--Associate Professor of Research,
\href{https://ouhsc.edu/bbmc/team/}{University of Oklahoma Health Sciences Center},
\href{https://medicine.ouhsc.edu/academic-departments/pediatrics}{Dept of Pediatrics},
Biomedical and Behavioral Methodology Core (\href{http://ouhsc.edu/bbmc/}{BBMC})

\href{https://www.vanderbilt.edu/psychological_sciences/bio/joe-rodgers}{Joseph Lee Rodgers}
--Director, Quantitative Methods, Department of Psychology and Human Development,
Peabody College, 230 Appleton Pl #552, Hobbs 202C,
Vanderbilt University, Nashville, TN 37203.

\href{https://www.slu.edu/medicine/health-and-clinical-outcomes-research/faculty/schuelke-matt.php}{Matthew Schuelke}
--Senior Statistician, Office of the Vice President for Research
Saint Louis University
}
\keyword{package}
