\name{AugmentCycleData}
\alias{AugmentCycleData}
\title{Shows the interrupted time series in Cartesian coordinates}
\usage{
AugmentYearDataWithMonthResolution(ds, dateName)
}
\arguments{
  \item{dsPlot}{The \code{data.frame} to cotaining the
  detailed data.}

  \item{xName}{The variable name in \code{dsPlot}
  containing the date}

  \item{linksPair}{The \code{data.frame} to validate.}
}
\value{
Returns a \code{ggplot2} graphing object
}
\description{
Shows the interrupted time series in Cartesian coordinates.
}
\examples{
a <- 32+323
}
\author{
Will Beasley
}
\keyword{linear}

