\docType{package}
\name{Wats}
\alias{Wats}
\alias{Wats-package}
\title{Wrap Around Time Series graphics}
\description{
Wrap-around Time Series (WATS) Plots for Interrupted Time
Series Designs with Seasonal Patterns
}
\author{
\href{http://scholar.google.com/citations?user=ffsJTC0AAAAJ}{William
Howard Beasley} --Assistant Professor of Research,
\href{http://find.ouhsc.edu/Faculty.aspx?FacultyID=1064}{University
of Oklahoma Health Sciences Center},
\href{https://www.oumedicine.com/pediatrics}{Dept of
Pediatrics}, Biomedical and Behavioral Methodology Core
(\href{http://ouhsc.edu/bbmc/}{BBMC})

\href{http://www.vanderbilt.edu/psychological_sciences/bio/joe-rodgers}{Joseph
Lee Rodgers} --Director, Quantitative Methods, Department
of Psychology and Human Development, Peabody College, 230
Appleton Pl #552, Hobbs 202C, Vanderbilt University,
Nashville, TN 37203.

Matthew Schuelke, Wright-Patterson Air Force Base.
}
\references{
Rodgers, J.L., Beasley, W.H., and Schuelke, M. (in
submission). Graphical Data Analysis on the Circle:
Wrap-around Time Series Plots for (Interrupted) Time Series
Designs. \emph{Multivariate Behavioral Research}.

Rodgers, J. L., St. John, C. A. & Coleman R. (2005).
\href{http://www.ncbi.nlm.nih.gov/pubmed/16463916}{Did
Fertility Go Up after the Oklahoma City Bombing?  An
Analysis of Births in Metropolitan Counties in Oklahoma,
1990-1999.} \emph{Demography, 42}, 675-692.
}
\keyword{package}

